[8054-REFACTO] - Created a `OrderedCollectionType` type 🥰🎉
The new OrderCollectionType will handle all complex logic related to ordering, creating, deleting and updating a collection within a form. For the moment, you have to explicitly set what field to update on the `on_update` options, but it could easily be changed to automatically calls the corresponding setters to only modified field between the collection and the db, but it is more than enough actually.

- Added support for SelectionStep

GitOrigin-RevId: 162e24ecf80071c1ea52060a89af4c5ad2794f51